fix: refactoring of textboxes view
