+++ views/ParametersView.py
@@ -2,7 +2,7 @@ from constants.Colors import Colors
 
 from models.TextBox import TextBox
 
-from views.TextBoxView import TextBoxView
+from views.text_boxes.TextBoxView import TextBoxView
 
 from controllers.text_boxes.TextBoxController import TextBoxController
 from controllers.buttons.ButtonController import ButtonController
